[WIP] Refactoring to better support unit tests.
